+++ Changelog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [UNRELEASED]
+### Added
+ - Multiple masks for text field and phone number components
+
 ## 3.0.0-alpha.5
 ### Fixed
  - Issues with the simple conditional logic in form builder.
@@ -11,7 +15,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## 3.0.0-alpha.4
 ### Added
  - Better documentation around the javascript execution code.
-
 ### Fixed
  - Issues with text area not working when set to required.
  - Problems where two text with same key not working in form builder.